+++ tapo-py/examples/tapo_l530.py
@@ -1,9 +1,10 @@
-"""L530, L630 & L900 Example"""
+"""L530 & L630 Example"""
 
 import asyncio
 import os
 
-from tapo import ApiClient, Color
+from tapo import ApiClient
+from tapo.requests import Color
 
 
 async def main():