+++ tapo/src/api/hub_handler.rs
@@ -7,8 +7,33 @@ use crate::api::{KE100Handler, S200BHandler, T100Handler, T110Handler, T300Handl
 use crate::error::Error;
 use crate::requests::TapoRequest;
 use crate::responses::{
-    ChildDeviceListResult, ChildDeviceResult, DeviceInfoHubResult, TapoResponseExt,
+    ChildDeviceHubResult, ChildDeviceListHubResult, DeviceInfoHubResult, TapoResponseExt,
 };
+
+macro_rules! get_device_id {
+    ($self:expr, $identifier:expr, $value:path) => {{
+        let children = $self.get_child_device_list().await?;
+
+        match $identifier {
+            HubDevice::ByDeviceId(device_id) => children
+                .iter()
+                .filter_map(|d| match d {
+                    $value(child) if child.device_id == device_id => Some(child.device_id.clone()),
+                    _ => None,
+                })
+                .next()
+                .ok_or_else(|| Error::DeviceNotFound)?,
+            HubDevice::ByNickname(nickname) => children
+                .iter()
+                .filter_map(|d| match d {
+                    $value(child) if child.nickname == nickname => Some(child.device_id.clone()),
+                    _ => None,
+                })
+                .next()
+                .ok_or_else(|| Error::DeviceNotFound)?,
+        }
+    }};
+}
 
 /// Handler for the [H100](https://www.tapo.com/en/search/?q=H100) hubs.
 pub struct HubHandler {
@@ -40,12 +65,13 @@ impl HubHandler {
         self.client.get_device_info().await
     }
 
-    /// Returns *child device list* as [`ChildDeviceListResult`].
+    /// Returns *child device list* as [`ChildDeviceHubResult`].
     /// It is not guaranteed to contain all the properties returned from the Tapo API
     /// or to support all the possible devices connected to the hub.
-    pub async fn get_child_device_list(&self) -> Result<Vec<ChildDeviceResult>, Error> {
+    /// If the deserialization fails, or if a property that you care about it's not present, try [`HubHandler::get_child_device_list_json`].
+    pub async fn get_child_device_list(&self) -> Result<Vec<ChildDeviceHubResult>, Error> {
         self.client
-            .get_child_device_list::<ChildDeviceListResult>()
+            .get_child_device_list::<ChildDeviceListHubResult>()
             .await
             .map(|r| r.devices)
     }
@@ -77,16 +103,16 @@ impl HubHandler {
 
 /// Child device handler builders.
 impl HubHandler {
-    /// Returns a [`S200BHandler`] for the given `device_id`.
+    /// Returns a [`KE100Handler`] for the given [`HubDevice`].
     ///
     /// # Arguments
     ///
-    /// * `device_id` - the Device ID of the child device
+    /// * `identifier` - a hub device identifier
     ///
     /// # Example
     ///
     /// ```rust,no_run
-    /// # use tapo::ApiClient;
+    /// # use tapo::{ApiClient, HubDevice};
     /// # #[tokio::main]
     /// # async fn main() -> Result<(), Box<dyn std::error::Error>> {
     /// // Connect to the hub
@@ -94,26 +120,28 @@ impl HubHandler {
     ///     .h100("192.168.1.100")
     ///     .await?;
     /// // Get a handler for the child device
-    /// let device = hub.s200b("0000000000000000000000000000000000000000");
+    /// let device_id = "0000000000000000000000000000000000000000".to_string();
+    /// let device = hub.ke100(HubDevice::ByDeviceId(device_id)).await?;
     /// // Get the device info of the child device
     /// let device_info = device.get_device_info().await?;
     /// # Ok(())
     /// # }
     /// ```
-    pub fn s200b(&self, device_id: impl Into<String>) -> S200BHandler {
-        S200BHandler::new(self, device_id.into())
+    pub async fn ke100(&self, identifier: HubDevice) -> Result<KE100Handler, Error> {
+        let device_id = get_device_id!(self, identifier, ChildDeviceHubResult::KE100);
+        Ok(KE100Handler::new(self, device_id))
     }
 
-    /// Returns a [`T100Handler`] for the given `device_id`.
+    /// Returns a [`S200BHandler`] for the given [`HubDevice`].
     ///
     /// # Arguments
     ///
-    /// * `device_id` - the Device ID of the child device
+    /// * `identifier` - a hub device identifier
     ///
     /// # Example
     ///
     /// ```rust,no_run
-    /// # use tapo::ApiClient;
+    /// # use tapo::{ApiClient, HubDevice};
     /// # #[tokio::main]
     /// # async fn main() -> Result<(), Box<dyn std::error::Error>> {
     /// // Connect to the hub
@@ -121,26 +149,28 @@ impl HubHandler {
     ///     .h100("192.168.1.100")
     ///     .await?;
     /// // Get a handler for the child device
-    /// let device = hub.t100("0000000000000000000000000000000000000000");
+    /// let device_id = "0000000000000000000000000000000000000000".to_string();
+    /// let device = hub.s200b(HubDevice::ByDeviceId(device_id)).await?;
     /// // Get the device info of the child device
     /// let device_info = device.get_device_info().await?;
     /// # Ok(())
     /// # }
     /// ```
-    pub fn t100(&self, device_id: impl Into<String>) -> T100Handler {
-        T100Handler::new(self, device_id.into())
+    pub async fn s200b(&self, identifier: HubDevice) -> Result<S200BHandler, Error> {
+        let device_id = get_device_id!(self, identifier, ChildDeviceHubResult::S200B);
+        Ok(S200BHandler::new(self, device_id))
     }
 
-    /// Returns a [`T110Handler`] for the given `device_id`.
+    /// Returns a [`T100Handler`] for the given [`HubDevice`].
     ///
     /// # Arguments
     ///
-    /// * `device_id` - the Device ID of the child device
+    /// * `identifier` - a hub device identifier
     ///
     /// # Example
     ///
     /// ```rust,no_run
-    /// # use tapo::ApiClient;
+    /// # use tapo::{ApiClient, HubDevice};
     /// # #[tokio::main]
     /// # async fn main() -> Result<(), Box<dyn std::error::Error>> {
     /// // Connect to the hub
@@ -148,26 +178,28 @@ impl HubHandler {
     ///     .h100("192.168.1.100")
     ///     .await?;
     /// // Get a handler for the child device
-    /// let device = hub.t110("0000000000000000000000000000000000000000");
+    /// let device_id = "0000000000000000000000000000000000000000".to_string();
+    /// let device = hub.t100(HubDevice::ByDeviceId(device_id)).await?;
     /// // Get the device info of the child device
     /// let device_info = device.get_device_info().await?;
     /// # Ok(())
     /// # }
     /// ```
-    pub fn t110(&self, device_id: impl Into<String>) -> T110Handler {
-        T110Handler::new(self, device_id.into())
+    pub async fn t100(&self, identifier: HubDevice) -> Result<T100Handler, Error> {
+        let device_id = get_device_id!(self, identifier, ChildDeviceHubResult::T100);
+        Ok(T100Handler::new(self, device_id))
     }
 
-    /// Returns a [`T300Handler`] for the given `device_id`.
+    /// Returns a [`T110Handler`] for the given [`HubDevice`].
     ///
     /// # Arguments
     ///
-    /// * `device_id` - the Device ID of the child device
+    /// * `identifier` - a hub device identifier
     ///
     /// # Example
     ///
     /// ```rust,no_run
-    /// # use tapo::ApiClient;
+    /// # use tapo::{ApiClient, HubDevice};
     /// # #[tokio::main]
     /// # async fn main() -> Result<(), Box<dyn std::error::Error>> {
     /// // Connect to the hub
@@ -175,26 +207,28 @@ impl HubHandler {
     ///     .h100("192.168.1.100")
     ///     .await?;
     /// // Get a handler for the child device
-    /// let device = hub.t300("0000000000000000000000000000000000000000");
+    /// let device_id = "0000000000000000000000000000000000000000".to_string();
+    /// let device = hub.t110(HubDevice::ByDeviceId(device_id)).await?;
     /// // Get the device info of the child device
     /// let device_info = device.get_device_info().await?;
     /// # Ok(())
     /// # }
     /// ```
-    pub fn t300(&self, device_id: impl Into<String>) -> T300Handler {
-        T300Handler::new(self, device_id.into())
+    pub async fn t110(&self, identifier: HubDevice) -> Result<T110Handler, Error> {
+        let device_id = get_device_id!(self, identifier, ChildDeviceHubResult::T110);
+        Ok(T110Handler::new(self, device_id))
     }
 
-    /// Returns a [`T31XHandler`] for the given `device_id`.
+    /// Returns a [`T300Handler`] for the given [`HubDevice`].
     ///
     /// # Arguments
     ///
-    /// * `device_id` - the Device ID of the child device
+    /// * `identifier` - a hub device identifier
     ///
     /// # Example
     ///
     /// ```rust,no_run
-    /// # use tapo::ApiClient;
+    /// # use tapo::{ApiClient, HubDevice};
     /// # #[tokio::main]
     /// # async fn main() -> Result<(), Box<dyn std::error::Error>> {
     /// // Connect to the hub
@@ -202,26 +236,28 @@ impl HubHandler {
     ///     .h100("192.168.1.100")
     ///     .await?;
     /// // Get a handler for the child device
-    /// let device = hub.t310("0000000000000000000000000000000000000000");
+    /// let device_id = "0000000000000000000000000000000000000000".to_string();
+    /// let device = hub.t300(HubDevice::ByDeviceId(device_id)).await?;
     /// // Get the device info of the child device
     /// let device_info = device.get_device_info().await?;
     /// # Ok(())
     /// # }
     /// ```
-    pub fn t310(&self, device_id: impl Into<String>) -> T31XHandler {
-        T31XHandler::new(self, device_id.into())
+    pub async fn t300(&self, identifier: HubDevice) -> Result<T300Handler, Error> {
+        let device_id = get_device_id!(self, identifier, ChildDeviceHubResult::T300);
+        Ok(T300Handler::new(self, device_id))
     }
 
-    /// Returns a [`T31XHandler`] for the given `device_id`.
+    /// Returns a [`T31XHandler`] for the given [`HubDevice`].
     ///
     /// # Arguments
     ///
-    /// * `device_id` - the Device ID of the child device
+    /// * `identifier` - a hub device identifier
     ///
     /// # Example
     ///
     /// ```rust,no_run
-    /// # use tapo::ApiClient;
+    /// # use tapo::{ApiClient, HubDevice};
     /// # #[tokio::main]
     /// # async fn main() -> Result<(), Box<dyn std::error::Error>> {
     /// // Connect to the hub
@@ -229,26 +265,28 @@ impl HubHandler {
     ///     .h100("192.168.1.100")
     ///     .await?;
     /// // Get a handler for the child device
-    /// let device = hub.t315("0000000000000000000000000000000000000000");
+    /// let device_id = "0000000000000000000000000000000000000000".to_string();
+    /// let device = hub.t310(HubDevice::ByDeviceId(device_id)).await?;
     /// // Get the device info of the child device
     /// let device_info = device.get_device_info().await?;
     /// # Ok(())
     /// # }
     /// ```
-    pub fn t315(&self, device_id: impl Into<String>) -> T31XHandler {
-        T31XHandler::new(self, device_id.into())
+    pub async fn t310(&self, identifier: HubDevice) -> Result<T31XHandler, Error> {
+        let device_id = get_device_id!(self, identifier, ChildDeviceHubResult::T310);
+        Ok(T31XHandler::new(self, device_id))
     }
 
-    /// Returns a [`KE100Handler`] for the given `device_id`.
+    /// Returns a [`T31XHandler`] for the given [`HubDevice`].
     ///
     /// # Arguments
     ///
-    /// * `device_id` - the Device ID of the child device
+    /// * `identifier` - a hub device identifier
     ///
     /// # Example
     ///
     /// ```rust,no_run
-    /// # use tapo::ApiClient;
+    /// # use tapo::{ApiClient, HubDevice};
     /// # #[tokio::main]
     /// # async fn main() -> Result<(), Box<dyn std::error::Error>> {
     /// // Connect to the hub
@@ -256,13 +294,23 @@ impl HubHandler {
     ///     .h100("192.168.1.100")
     ///     .await?;
     /// // Get a handler for the child device
-    /// let device = hub.ke100("0000000000000000000000000000000000000000");
+    /// let device_id = "0000000000000000000000000000000000000000".to_string();
+    /// let device = hub.t315(HubDevice::ByDeviceId(device_id)).await?;
     /// // Get the device info of the child device
     /// let device_info = device.get_device_info().await?;
     /// # Ok(())
     /// # }
     /// ```
-    pub fn ke100(&self, device_id: impl Into<String>) -> KE100Handler {
-        KE100Handler::new(self, device_id.into())
+    pub async fn t315(&self, identifier: HubDevice) -> Result<T31XHandler, Error> {
+        let device_id = get_device_id!(self, identifier, ChildDeviceHubResult::T315);
+        Ok(T31XHandler::new(self, device_id))
     }
+}
+
+/// Hub Device.
+pub enum HubDevice {
+    /// By Device ID.
+    ByDeviceId(String),
+    /// By Nickname.
+    ByNickname(String),
 }