+++ tapo/src/requests/get_energy_data.rs
@@ -11,17 +11,34 @@ pub(crate) struct GetEnergyDataParams {
 
 impl GetEnergyDataParams {
     pub fn new(interval: EnergyDataInterval) -> Self {
+        let timezone = chrono::Local::now().timezone();
+
         match interval {
             EnergyDataInterval::Hourly {
                 start_date,
                 end_date,
             } => Self {
-                start_timestamp: start_date.and_hms_opt(0, 0, 0).unwrap().timestamp() as u64,
-                end_timestamp: end_date.and_hms_opt(23, 59, 59).unwrap().timestamp() as u64,
+                start_timestamp: start_date
+                    .and_hms_opt(0, 0, 0)
+                    .unwrap()
+                    .and_local_timezone(timezone)
+                    .unwrap()
+                    .timestamp() as u64,
+                end_timestamp: end_date
+                    .and_hms_opt(23, 59, 59)
+                    .unwrap()
+                    .and_local_timezone(timezone)
+                    .unwrap()
+                    .timestamp() as u64,
                 interval: 60,
             },
             EnergyDataInterval::Daily { start_date } => {
-                let timestamp = start_date.and_hms_opt(0, 0, 0).unwrap().timestamp() as u64;
+                let timestamp = start_date
+                    .and_hms_opt(0, 0, 0)
+                    .unwrap()
+                    .and_local_timezone(timezone)
+                    .unwrap()
+                    .timestamp() as u64;
                 Self {
                     start_timestamp: timestamp,
                     end_timestamp: timestamp,
@@ -29,7 +46,12 @@ impl GetEnergyDataParams {
                 }
             }
             EnergyDataInterval::Monthly { start_date } => {
-                let timestamp = start_date.and_hms_opt(0, 0, 0).unwrap().timestamp() as u64;
+                let timestamp = start_date
+                    .and_hms_opt(0, 0, 0)
+                    .unwrap()
+                    .and_local_timezone(timezone)
+                    .unwrap()
+                    .timestamp() as u64;
                 Self {
                     start_timestamp: timestamp,
                     end_timestamp: timestamp,