+++ CHANGELOG.md
@@ -6,12 +6,81 @@ file. This change log follows the conventions of
 
 ## [Rust Unreleased][Unreleased]
 
+## [Python Unreleased][Unreleased]
+
+## [Rust v0.7.12][v0.7.12] - 2024-06-27
+
+### Changed
+
+- H100's create child device handler methods now take a `HubDevice` `enum` instead of a `String` and are now `async` to allow for more flexibility. This enables the caller to find child devices by either device ID or nickname.
+- `PlugIdentifier` has been renamed to `Plug`.
+- `Plug::ByDeviceId` now verifies that the provided device ID is found and returns an `Error::DeviceNotFound` error when it's not.
+- `HubDevice` variants now take a `String` instead of a `&str` to allow for more flexibility.
+- `Plug` variants now take a `String` instead of a `&str` to allow for more flexibility.
+
+### Fixed
+
+- `ColorLightSetDeviceInfoParams` `hue` field validation has been changed from `between 1 and 360` to `between 0 and 360` to match the device's expected range.
+- Fixed an issue where the `EnergyDataResult's `start_timestamp` and `end_timestamp` did not correctly adjust for timezone offsets.
+- The `chrono` dependency has been updated to `0.4.34` to fix the minimum version requirement.
+
+### Removed
+
+- The `overheated` property has been removed from `DeviceInfoGenericResult` because it's not present in the response of all devices.
+
+## [Python v0.3.1][py-v0.3.1] - 2024-06-27
+
+### Fixed
+
+- `ColorLightSetDeviceInfoParams` `hue` field validation has been changed from `between 1 and 360` to `between 0 and 360` to match the device's expected range.
+- Fixed an issue where the `EnergyDataResult's `start_timestamp` and `end_timestamp` did not correctly adjust for timezone offsets.
+- All handlers are now correctly exported and can be imported from the `tapo` module.
+
+### Removed
+
+- The `overheated` property has been removed from `DeviceInfoGenericResult` because it's not present in the response of all devices.
+
+## [Rust v0.7.11][v0.7.11] - 2024-05-04
+
+### Added
+
+- Added support for the P300 power strip (thanks to @Michal-Szczepaniak).
+- `RgbLightStripHandler` and `DeviceInfoRgbLightStripResult` have been added to support the L900 devices separately from the L530 and L630 devices.
+
+### Changed
+
+- `ChildDeviceResult` has been renamed to `ChildDeviceHubResult` to facilitate adding support for other devices with children.
+- `ColorLightStripHandler` has been renamed to `RgbicLightStripHandler` to better reflect its purpose.
+- `DeviceInfoColorLightStripResult` has been renamed to `DeviceInfoRgbicLightStripResult` to better reflect its purpose.
+
+## [Python v0.3.0][py-v0.3.0] - 2024-05-04
+
+### Added
+
+- Added partial support for the H100 hub and its child devices. Currently, only the `get_device_info` function is supported for the child devices through the hub's `get_child_device_list` method.
+
+### Changed
+
+- A large number of types have been reorganized to me more in line with the Rust library. This includes moving many of them under the `requests` and `responses` sub modules.
+
+### Removed
+
+- `l900` has been removed from the `ApiClient` until proper support is added.
+
+## [Rust v0.7.10][v0.7.10] - 2024-04-05
+
 ### Changed
 
 - The implementation of `ApiClient::new` has been improved to allow for the return of `ApiClient` instead of `Result<ApiClient, Error>`.
 - The default timeout for all requests has been reduced to 30 seconds from 300 seconds.
+- `ApiClient::with_timeout` has been added to allow for the setting of a custom timeout for all requests (thanks to @skoky).
 
-## [Python Unreleased][Unreleased]
+## [Python v0.2.1][py-v0.2.1] - 2024-04-05
+
+### Changed
+
+- The default timeout for all requests has been reduced to 30 seconds from 300 seconds.
+- The `timeout_s` optional parameter has been added to the `ApiClient` constructor to allow for the setting of a custom timeout for all requests (thanks to @skoky).
 
 ## [Rust v0.7.9][v0.7.9] - 2024-01-27
 
@@ -297,6 +366,12 @@ let device = ApiClient::new(ip_address, tapo_username, tapo_password)?
 ### Initial Release of Tapo
 
 [Unreleased]: https://github.com/mihai-dinculescu/tapo
+[v0.7.12]: https://github.com/mihai-dinculescu/tapo/tree/v0.7.12
+[py-v0.3.1]: https://github.com/mihai-dinculescu/tapo/tree/py-v0.3.1
+[v0.7.11]: https://github.com/mihai-dinculescu/tapo/tree/v0.7.11
+[py-v0.3.0]: https://github.com/mihai-dinculescu/tapo/tree/py-v0.3.0
+[v0.7.10]: https://github.com/mihai-dinculescu/tapo/tree/v0.7.10
+[py-v0.2.1]: https://github.com/mihai-dinculescu/tapo/tree/py-v0.2.1
 [v0.7.9]: https://github.com/mihai-dinculescu/tapo/tree/v0.7.9
 [py-v0.2.0]: https://github.com/mihai-dinculescu/tapo/tree/py-v0.2.0
 [v0.7.8]: https://github.com/mihai-dinculescu/tapo/tree/v0.7.8